refactor(flow-core): centralize describeFinding to end per-surface drift
The org-health finding renderer had drifted into four near-identical copies
(CLI audit, CLI monitor, GUI, extension) that disagreed on field handling —
licenses emit 'total' while limits emit 'max', and only the CLI monitor copy
rendered the limit ratio %. That divergence was the root of the earlier
license-rendering bug.

- Add packages/flow-core/src/health-findings.ts: one describeFinding() that is
  the union of all four — apiVersion / user / audit-trail / apex-job / license
  (total) / limit (max + ratio%) / health-score / backup-error shapes — exported
  from the flow-core barrel.
- Delete the four local copies; CLI (audit.js, monitor.js), gui (HealthChecks.jsx)
  and extension (org-health.ts) now import it from @sfdt/flow-core.
- Add @sfdt/flow-core as a dependency of @sfdt/gui (Vite + vitest resolve it from
  the workspace; production build verified).

flow-core stays a separate, zero-dependency, browser-safe package precisely so
the Chrome extension can share this logic without pulling in the Node-only CLI.

Verified: lint clean, extension tsc compile, gui build, 2020/2020 tests pass.
