+++ gui/package.json
@@ -11,6 +11,7 @@
     "test:watch": "vitest"
   },
   "dependencies": {
+    "@sfdt/flow-core": "^0.9.0",
     "d3": "^7.9.0",
     "react": "^19.2.7",
     "react-dom": "^19.2.7",
